requirements.txt
fastapi~=0.110.0
python-dotenv~=1.0.1
starlette~=0.36.3

coverage~=7.4.1
httpx~=0.27.0
pytest~=8.1.1
pytest-html~=4.1.1
mypy~=1.9.0
tox~=4.14.0
